requirements.txt
pydantic>=2.12.0

# API server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
requests>=2.31.0

# Orchestration
# NOTE: Keep versions pinned to avoid Python 3.12 forward-ref issues.
langgraph==0.0.26
langchain-core==0.1.52
langsmith==0.1.81

# LLM client + data stores
groq>=1.0.0
redis>=5.0.0
pymongo>=4.6.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0

# ETL utilities (optional but used by scripts)
pandas>=2.2.0

# Testing
pytest>=8.0.0

# UI
streamlit>=1.33.0
